create resuable bookmark button for videos
