+++ app/src/components/VideoCard/VideoCard.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from "react";
 import emptyBookMark from "../../../public/assets/icon-bookmark-empty.svg";
 import movie from "../../../public/assets/icon-category-movie.svg";
 import tv from "../../../public/assets/icon-category-tv.svg";
+import BookMarkButton from "../BookMarkButton/BookMarkButton";
 import "./VideoCard.scss";
 
 import { useBookMark } from "../../hooks/useBookMark";
@@ -16,24 +17,7 @@ function VideoCard({
   url,
   year,
 }) {
-  // state for when a video is bookmarked
-  const [isBookmarked, setBookmark] = useState(false);
-
-  const { addBookMark, deleteBookMarkVideo } = useBookMark();
-  const handleBookmark = async () => {
-    await addBookMark(id);
-  };
-
-  const removeBookMark = async () => {
-    await deleteBookMarkVideo(id);
-    console.log("clicked");
-  };
-
-  useEffect(() => {
-    if (is_bookmarked) {
-      setBookmark(true);
-    }
-  }, []);
+  
 
   return (
     <>
@@ -46,29 +30,13 @@ function VideoCard({
             backgroundPosition: "center",
           }}
         >
-          {isBookmarked ? (
-            <button
-              className="videoContainer__bookmarked"
-              onClick={removeBookMark}
-            >
-              <img
-                className="videoContainer__bookmark-icon"
-                src={emptyBookMark}
-                alt="bookmark icon"
-              />
-            </button>
-          ) : (
-            <button
-              className="videoContainer__book-btn"
-              onClick={handleBookmark}
-            >
-              <img
-                className="videoContainer__bookmark-icon"
-                src={emptyBookMark}
-                alt="bookmark icon"
-              />
-            </button>
-          )}
+          <BookMarkButton
+          is_bookmarked={is_bookmarked}
+          prefixClass={"videoContainer"}
+          bookMarkIcon={emptyBookMark}
+          videoId={id}
+          />
+        
         </div>
         <div className="videoContainer__content">
           <div className="videoContainer__text-content">
